Removed the superfluous files and directories and added to the .gitignore
Revert back the config for development as it was used for testing and
should be set back.

Deleted the generate manifest.py and used the Jinja2 command to be
similar to other repos.

Removed a few more PyCharm files

Moved to the Makefile similar to dd-api so the two are more consistent
which includes a few associated files.

A few more tweaks to make this more inline with the dd-api

Stripped out some more superfluous stuff as it wasn't needed and was
just extra to maintain.