use FLASK_ENV instead of ENVIRONMENT env var
flask needs the environment to be set before it starts (ie - not set
in code). It expects the FLASK_ENV variable to populated. This commit
sets that in the makefile, and removes references to the environment
variable. If you need to access the env in code, you can either look at
`os.environ('FLASK_ENV')` or `current_app.env`.

This is preferable, because if it isn't set then flask defaults to
production settings - ie: debug turned off. Some plugins might read
from this rather than from our own settings too

http://flask.pocoo.org/docs/1.0/config/#environment-and-debug-features