Remove all Javascript
We don’t serve any Javascript to the user, so our build pipeline no
longer needs to do anything Javascripty.

This reduces the number of dependencies we need to track.

I don’t really know how the `javascripts` task worked in the first place
since it was missing a bunch of the source files it was supposed to be
compiling.