enable running via docker locally
on paas it just runs through gunicorn.

locally to run through docker it needs to be able to talk to api (by
using host.docker.internal, which represents your machine's localhost).
also, it needs to broadcast on 0.0.0.0 to allow things outside the
docker container (like your browser) to contact it

also remove a bunch of unused env vars from makefile