Fix tests
Fix 1.

The find method was called looking for an anchor
with the exact contents of "Continue".

The new version of that anchor includes some
whitespace around "Continue".

Fix 2.

UTF8 encoding was removed from the asset
fingerprinter to allow it to process images. This
moves that to the tests instead, to copy the
approach for similar tests in the admin app:

https://github.com/alphagov/notifications-admin/commit/643dca0071e52d6aaca5a87384beedb29fb6e92b