Showing an error page if a file is not in S3
If a file was not in S3, the user would see an S3 error when
clicking on the link to download the file. document-download-api now has
an endpoint to check if a file is in S3, which means that when a user
clicks an email link, we can check if the file is available and only
show the page to download it if it is.

If there was an unexpected error when checking if the file was
available, the error handlers will handle it and document-download-api
logs a message so we can investigate.