Show 404 page when visiting link with wrong encryption key
We had accounted for the case where someone visits a link with no key
or a key that can't be decoded using `urlsafe_b64decode`. However, if
there is a key and that key can be decoded using `urlsafe_b64decode`,
the error returned to document-download-api by S3 is a `Forbidden`
error which wasn't being taken into account.

If S3 gives a `Forbidden` error, because the key is wrong, we now show
the `404` page. This page already contains instructions about checking
the link.

The JSON returned from document-download-api's endpoint to check if a
file is in S3 contains the key `error`, not `Error`, so this also fixes
that.