Use URL from API for direct doc download
Previously we built this manually. Using the URL provided by the
API reduces coupling between the two apps, and will automatically
mean we start using the new URL that includes a file extension,
which this app didn't have the data for.

In the next commits we'll look at how we deal with failure, since
we're now calling the API in both endpoints.