Favour end-to-end testing over method stub
Previously we tested and stubbed the behaviour of method to get
document metadata, which held a greater risk that the behaviour
of the method would start to diverge from reality. This switches
to an end-to-end testing approach by creating fixtures to allow
us to stub the actual API call in each test, DRYing up some of
the setup for each test along the way.

Note that we leave the stubbing as-is for the errorhandler test,
since the choice of endpoint to test here is arbitrary, so:

- The test doesn't benefit from an end-to-end approach
- It doesn't justify defining the new fixtures at the top level