Increase memory from 512MB to 1G
This app has been skirting around the 95% memory mark for the last 2
months. There is a chance this is causing the occasional error. We don't
have proper evidence for this but it is a hypothesis that we can quickly
test.

It may be benficial to bump the memory regardless of whether it fixes
occasional errors, just to give the app some more headroom in case it's
memory requirements go up very slightly in the future.