Add security.txt page
This follows the guidance in
https://gds-way.cloudapps.digital/standards/vulnerability-disclosure.html#vulnerability-disclosure-and-security-txt

Note, I plan on just adding this for www.notifications.service.gov.uk
and documents.service.gov.uk.

I think we can get away with not adding it to the tech-docs or the
API but happy to be challenged on that

See https://github.com/alphagov/notifications-admin/pull/4088 for
equivalent on the admin app.