Add package-lock.json, use Node 16.14.0 LTS & NPM 8.3.1
Bumps Node version required to build the project to 16.14.0 - the current
LTS version this comes with NPM 8.3.1, but it's also safe to use a later
minor version.

Note the use of caret `^` in the engine version definiton. This means
we can continue to build using minor and patch updates to Node 16 but
will not support building using major versions beyond the currrent LTS version.
This will keep parity with the CI build environment.

Adds package-lock.json to ensure consistent package versions are installed
at build time and in local development environments.

Signed-off-by: Richard Baker <richard.baker@digital.cabinet-office.gov.uk>
Co-authored-by: Tom Byers <tom.byers@digital.cabinet-office.gov.uk>