Change README to advise using nvm instead of n
We want a single way to set the version of NodeJS
for this project so it is consistent across
environments and easy to change.

There are many different version managers for
NodeJS and no real standard for specifying a
version of NodeJS. In absence of any industry
standards, we polled the devs on Notify and those
in the GDS frontend and NodeJS communities on
which version manager they use. NVM was
consistently the most used tool for this and, when
asked, the Notify devs agreed they would be happy
switching to it.

This updates the README with instructions on how
to use NVM rather than n.