Remove `engines` part of package.json
We were using the `engines` part of package.json
as the one place to set the version of NodeJS to
use. Now we're using a `.nvmrc` file for this, I
don't think we need to set `engines` any more and
I can't see any other process that relies on it so
this removes it.