Update JS build to match govuk alerts
For some reason, the code we will need to add to
the govuk-frontend-local.mjs file breaks the JS
build. The error raised is that the
nodeListForEach function isn't exported by GOVUK
Frontend but, looking at the module this
references, it is.

Changing the JS build to match the one in
notifications-govuk-alerts fixes this. Those
changes include:
- using the JS API of rollupJS directly, not
  through the gulp-better-rollup plugin
- rollupJS and its plugins are bumped to latest
  versions
- we use terser to compress the output, not
  uglifyJS

Guesses at which ones of these fixed it are:
- removing gulp-better-rollup because it hasn't
  been updated in 3 years and so uses an older
  version of rollupJS
- an older version of rollupJS or one of the
  plugins we use to manage module imports had an
  issue

Fix issues in gulpfile