Fix how skiplinks & buttons initialise
As pointed out in this comment on the associated
pull request, there should only ever be one
skip-link per page so we don't need to iterate
over a nodelist. Also that the buttons are being
initialised incorrecty:

https://github.com/alphagov/document-download-frontend/pull/117/files/97996bd6d681e8c61255b52eb1f0510da5293e53..1996f79c38538cdf4e4c26ee2aefedd860865a25#r857845017

The current code for the buttons attaches an event
handler to the document rather than each node that
has the `govuk-button` data attribute. It still
worked because its callback guards against
elements without a role of 'button' but this is an
accident rather than intentional.

These changes:
- assume one skip-link per page
- look for all buttons in the page and create an
  instance of Button for each, mirrroring how
  GOVUK Frontend do it

See the GOVUK Frontend code for examples of the
last point:

https://github.com/alphagov/govuk-frontend/blob/main/src/govuk/all.js