Add better-npm-audit to check production packages
This will be run using a periodic Concourse job to notify us about
any new vulnerabilities in packages.

"npm audit" doesn't support ignoring issues, which may be necessary
if a fix isn't (currently) available. To get around this we use a
wrapper called better-npm-audit, which support ignoring individual
CVEs with an expiry [^1]. I've added an ignore file now so there's
something to point to in our docs.

Note: we don't want to run this as part of the build as this could
prevent us shipping priority changes. Although we should be fixing
CVEs, they are seldom urgent and can be deferred as long as we have
a process in place to address them.

[^1]: https://www.npmjs.com/package/better-npm-audit