Add further protection for leaking of urls by `Referer` header
In https://github.com/alphagov/document-download-frontend/pull/121
we added the `Referrer-Policy` header to help protect against
older browser versions from sending the url of the page using the
`Referer` header. However it turns out `Referrer-Policy` isn't
supported by all browsers, only roughly 95%:
https://caniuse.com/referrer-policy

There are some additional things we can do though to target
a slightly larger number of browsers.

1. `<meta name="referrer" content="never">`
This is the original draft spec for `Referrer-Policy` and is
supported by a few more browser versions than the current spec.
Read more about it in [1] on https://caniuse.com/referrer-policy.
This combined with the previous pull request is supported by
about 96% of browsers.

2. `rel="noreferrer"`.
Supported by about 96% of browsers: https://caniuse.com/rel-noreferrer
Can be applied on a link per link basis.
Has not been applied to the links to the national archives at the
bottom of the page because these are contained deep in the GOV.UK
Frontend and I don't think it is proportionate to duplicate all
of that code just to add the `rel="noreferrer"` to them.

Both of the above measures may give us an extra percent or so of
protection but we can't guarantee they will completely protect us
from users that support the `Referer` header but don't support any
of the above methods.