bump govuk-frontend-jinja to match what we require in package.json
We have GOV.UK Frontend as a dependency in two places:

* In python, our requirements.in specifies a version of
  [govuk-frontend-jinja](https://github.com/LandRegistry/govuk-frontend-jinja)
  for our jinja templates
* In node, our package.json specifies a version of
  [govuk-frontend](https://github.com/alphagov/govuk-frontend) for our
  fonts, images and sass files

We need to ensure that the version of govuk-frontend that the python
library relies on always matches the version of govuk-frontend in our
package.json exactly.

Bump the python version so that it matches the sass etc from our
package.json, and then add some comments in the README.md so we know to
do this again in the future.