Remove links to nationalarchives.gov.uk from the footer
This was causing us issues as some older browsers would send the URL the
user is on to nationalarchives.gov.uk via the HTTP referer header. We've
reduced the chance of this happening in a couple of pull
requests[^1][^2] but some browsers will still send this data across. As
admins on the nationalarchives could potentially see these URLs in their
access logs, it's best not to send them across.

To fix this, we simply copied the footer HTML across from the
govuk-frontend-jinja repo, and updated it to remove the links to the
nationalarchives.gov.uk website. If the footer changes in a future
version of govuk-frontend we should be careful to ensure we update this
file and don't undo our changes.

---

[^1]: https://github.com/alphagov/document-download-frontend/pull/121
[^2]: https://github.com/alphagov/document-download-frontend/pull/122