Add form for confirming email address
The form to confirm an email address contains a single input, the email
address field. It currently validates that the field is not empty and
that the email address is valid (using our validation rules from
notifications-utils).

This is currently expected to be the only form in the app, so I've gone
for a simpler approach than the one we have in admin, where fields and
widgets need to be used on multiple forms. We can hardcode some details
in the widget and define it within the `EmailAddressField` since it will
only be used by that field.

The widget for the `EmailAddressField` needs to return customisable
markup for the `govukInput` component. Since the template for the
component isn't exposed by GOV.UK Frontend, we have to create a file
which just calls the macro with params we can pass in for the widget to
use.