Handle CSRF errors
When CSRF validation fails, it will raise a `CSRFError` with a reason
and `400` status code. This copies the approach we have in admin of
logging the reason and displaying the `500` error page instead since the
error message is more appropriate to this situation - the heading of the
`400` error page is "Page not found".

Some fixtures needed moving so that they can be used in other files.