Add pre-commit hooks
We can offer pre-commit hooks to developers so that it's not possible
to commit code that breaks some basic formatting conventions.

The provided hooks for flake8 and isort will also fix issues
automatically, but reject the initial commit. Generally you can just add
the changes that the hooks have made, and then commit again, to accept
the automated fixes.

This change will do nothing by default - developers will need to opt-in
by installing the hooks.