Add logging to understand error better
In the latest release, a small percentage of requests raised KeyErrors
on missing `verify_email` keys. We believe this key should always be
present.

This patch does a soft retrieve of `verify_email` so that it won't error
if the key is missing, and logs the document metadata if it's missing so
that we can get some more information on what might be happening.