Show pretty file type on the download link for a file
https://www.pivotaltracker.com/story/show/183012123

Karls initial suggestion was that we should handle any unexpected
file extensions that aren't covered by our dictionary by referring
to them as their file extension. However I think I'd prefer that we
fail hard and get an error to indicate that we have missed something
rather than failing soft and not knowing that we've missed something.
When we add support for other filetypes, we will need to remember to
change the frontend too.