Return "file unavailable" if past expiry date
We have some unusual behaviour are files at the moment:

1) If the file isn't present in S3, we show a "file unavailable" page.
2) If the file is available but the custom retention period has passed,
we serve a 404.

This can result in a user seeing a 404 page (which doesn't have service
contact info) if they try to access the file soon after the expiry date
has passed (when the file hasn't been deleted from S3) yet.

Let's always serve the "file unavailable" page in this situation.