Make ECMAScript Module imports bare
Changes all imports of ECMAScript Modules in
`govuk-frontend-local.mjs` to be 'bare' (not
specifying a path and instead relying on the
module loader to source it) to make them less
prone to internal changes in the 'govuk-frontend'
package.

Also includes a cut-and-paste of the
`nodeListForEach` function now as the
'govuk-frontend' package doesn't export it.