Add 'none' to frame-src in CSP
`frame-src` was removed because there was nothing in it (after Youtube
was removed). However, this means that it will fall back to the
`default-src` and have a value of `'self' 'unsafe-inline'` which isn't
what we want and makes things less restrictive.

We now give it a value of `'none'` since we don't want any iframes used.