Allow cookie to be sent to subdomain if necessary
To confirm a user's email address we set a cookie at the frontend which
is then need to be passed to the api in order for it to be validated.

This works fine when both the frontend and the api are under the same
domain but when we're in a subdomain, we need to explicitly allow the
cookie to be sent to subdomains too.

Note that we could set the domain attribute of the cookie to
".documents.notify.works" (note the "." prefix) regardless of whether or
not the api is under a different domain, but this is safer. Given that
we control the entire domain I don't think it would be much of a risk,
but why do it when there's no reason to?

Docs: https://developer.mozilla.org/en-US/docs/Web/HTTP/Cookies#domain_attribute

Co-Authored-By: Katie Smith <katie.smith@digital.cabinet-office.gov.uk>