Split doc download API into external and internal routes
This means that internal traffic can be sent to the
internal URL for doc download API and references to the external
web address for doc download API can also be used in our code.

This won't change any behaviour on the PaaS as we set the
internal and external hosts to be the same.