Update `ServiceApiClient` to add new header
We want to pass the `X-Custom-Forwarder` header through in addition to
the three we were using because without it requests from this app to the
api are resulting in a 403 status code on ECS. This keeps things
consistent between the PaaS and ECS apps for now, but when everything is
migrated we can remove the need to check for the header in the apps.

To do this, the `ServiceApiClient` has been modified to inherit from the
`BaseAPIClient` instead of just using it. This means that we can
override the `generate_headers` method to add in `X-Custom-Forwarder`.