download FE unit tests to run in the doc download FE docker image, not the generic unit-test-image
When running the tests as part of the ECS build-document-download-frontend-main job we want to use the Dockerfile that the app uses instead of the generic test image. The app uses this Dockerfile when running so we want to be able to spot any issues when the code runs with it.

This line was removed because nvm is no longer used

We also upgraded the base image of the container to bookworm.

The Dockerfile was split in BASE, PRODUCTION, TEST.

The images build and the tests were run, although we need to make sure to pause STAGING AND PRODUCTION pipelines, as they will fail the tests.

This change also caused the notifications-aws to have its document-download-frontend pipeline upgraded to have the test image being in the pipeline