Add route proxy header to requests made to Notify API
Requests made to the Notify API that don't go through cloudfront,
require a special proxy header to be added. This proves that
they are legitimate requests from us coming in to a internal domain
name, rather than someone external finding one of our internal
domain names and using them.

The previous commit gets doc download FE calling the Notify API
on its internal domain name. Therefore, to suppport this, we also
need to add the proxy header to the API calls we make to the Notify
API. We already do this for the admin app making calls to the
Notify API:
https://github.com/alphagov/notifications-admin/blob/main/app/notify_client/__init__.py#L30

We have to add the route secret into the PaaS manifest to support
this. We don't need to add any secrets into ECS because we do not
have the special header checking in ECS (it is turned off):
https://github.com/alphagov/notifications-aws/blob/main/terraform/notify-apps/service_api.tf#L84