Add help text to all pages and make consistent
Following and prototyping session we decided to
just copy the help text from the last page without
any additional formatting (like inset text), so as
to introduce as little visual noise as possible.

These changes are to help the app comply with the
consistent help success criterion (3.2.6) from
WCAG 2.2 which asks us to put help in the same
place across multiple pages of a process.