Introduce a Rollup config
Config contains two separate sections:

1. ESM JavaScript bundle compilation
2. SCSS compilation

1. ESM JavaScript bundle compilation

Single input to create
a minified ESM bundle that contains links to
sourcemaps.

As Rollup "task" requires and input and output,
and the copy plugin doesn't have separate ones,
it's been added to this one.

2. SCSS compilation

Sadly, globbing isn't supported so we need to
list the entry points we import.

We minimise include links to sourcemaps and we
disable the CSS URL resolver
warnings (as referenced static files exists
elsewhere). For production build we also rewrite
the url to those assets as they're stored elsewhere.