Update Rollup configuration to support rebrand
Rebrand work brings in new static assets (favicons,
et al.) that lives in the `assets/rebrand` folder
of the package.
https://github.com/alphagov/govuk-frontend/releases/tag/v5.10.0#:~:text=Add%20the%20brand%20refresh%20assets%20to%20your%20project

Until the rebrand launch date, we still need to serve
current assets and then be able to quickly
update on the day.

To support this, the config has been tweaked with
- an `enableRebrand` Boolean which if true
will copy assets from the `assets/rebrand` folder
instead of the `assets/` one
- a new `govukFrontendAssetPaths` config object
where we set the paths for above assets based on the
`enableRebrand` Boolean.

To switch to the rebrand assets, we need to set
`enableRebrand: true`, merge and deploy.