Dockerfile: Inline contents of unit-tests-image
We want to migrate all of our images from GHCR to ECR, and as part of
that we want to remove/retire any images that we no longer need (rather
than migrating them to ECR and continuing to build/use them).

unit-tests-image is currently used as a base image for both admin &
document-download-frontend. This is not ideal because (presumably) both
apps have different dependency requirements, but by sharing the same
base image they both have the union of each of their requirements, which
is a waste of storage.

This commit inlines the contents of unit-tests-image's Dockerfile into
document-download-frontend's.