Dockerfile: Base frontend_build on node image (only)
Previously this image was based on a Python image which had the
/usr/local/bin and /usr/local/lib directories copy-pasted into it from a
Node image.

This is brittle because the Node and Python images might be referencing
different versions of Debian. It's also unnecessary because we don't use
Python at all during frontend builds.

Also reference Debian bookworm specifically, to ensure we're using the
same versions of Debian between the base image and the frontend_build
image.